+++ tcs/sam_mw_pt_type222_csp_solver.cpp
@@ -238,7 +238,7 @@ public:
 		mspt_receiver.m_th_tube = value(P_th_tu);				//[mm] Thickness of receiver tubes
 
 		mspt_receiver.m_flow_type = (int)value(P_Flow_type);	//[-] Numerical code to designate receiver flow type
-        mspt_receiver.m_crossover_shift = value(P_crossover_shift);
+		mspt_receiver.m_crossover_shift = (int)value(P_crossover_shift);
 
 		mspt_receiver.m_epsilon = value(P_epsilon);				//[-] Emissivity of receiver
 		mspt_receiver.m_hl_ffact = value(P_hl_ffact);			//[-] Heat Loss Fudge FACTor
@@ -273,7 +273,7 @@ public:
 		// Are we modeling a direct ISCC case?
 		mspt_receiver.m_is_iscc = value(P_IS_DIRECT_ISCC) == 1;
 		// Set cycle configuration in class
-		mspt_receiver.m_cycle_config = value(P_CYCLE_CONFIG);
+		mspt_receiver.m_cycle_config = (int)value(P_CYCLE_CONFIG);
